+++ client/index.html
@@ -25,10 +25,6 @@
     ></script>
     <script src="https://cdn.jsdelivr.net/npm/swiper@8/swiper-bundle.min.js"></script>
     <script type="module" src="./js/addCart.js"></script>
-    <script
-      type="module"
-      src="./js/addCartPopupCount.js"
-    ></script>
     <script defer src="./js/swiper.js"></script>
     <script defer src="./js/popup.js"></script>
     <script type="module" src="./js/main.js"></script>
